Update workflow to create PRs for DEV and QA stages
Added steps to 'create-pr-on-commit-to-main.yml' to automatically create Pull Requests when a commit is made to the main branch. These PRs propose to merge the changes to DEV and QA stages. This change streamlines the process of propagating updates from the main branch to the staging environments. This ensures that the DEV and QA stages always reflect the latest approved changes.