Remove delete-branch option in Github Actions
The 'delete-branch' option was removed from the Pull Request (PR) creation workflow in 'create-pr-on-commit-to-main.yml'. This avoids automatic deletion of the update-stages/dev and update-stages/qa branches after the PR merges. By preserving these branches, it allows us to refer back to the branch, as needed, for further changes or reference.