Update PR creation workflow in Github Actions
Changes were made to the 'create-pr-on-commit-to-main.yml' workflow file. Notify, the 'branch' option under the uses of 'peter-evans/create-pull-request' was changed from 'update-stages/dev' and 'update-stages/qa' to 'main' for both the DEV and QA stages. This change was made to ensure PRs are created on the main branch henceforward, streamlining and consolidating all the changes on a single branch.