Add 'git pull' step in Github Actions workflow to update QA
This commit introduces 'git pull' in the Github Actions workflow in the 'update-qa-on-commit-to-main.yml' file. This command will ensure the merging process fetches not just the metadata but also the changes from the 'main' branch. As a result, QA will have an up-to-date version of the main stage at any given point.