Update Github Actions workflow to keep QA in sync with main
This commit updates the workflow in 'update-qa-on-commit-to-main.yml'. The name of the workflow is changed from "Add changes to DEV stage" to "Add changes to QA stage". This is done to more accurately reflect the purpose of the workflow and ensure QA environment is up-to-date with latest main branch changes.